requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
boto3==1.35.60
botocore==1.35.60
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
datasets==3.1.0
dill==0.3.8
distro==1.9.0
faiss-cpu==1.9.0
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langchain-together==0.2.0
langsmith==0.1.143
MarkupSafe==3.0.2
marshmallow==3.23.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.17.1
